Init migration and seed default user.

+++ Scheduler.Domain/Entities/Attendee.cs
@@ -3,10 +3,10 @@
     public class Attendee : IEntityBase
     {
         public long Id { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
 
-        public int ScheduleId { get; set; }
+        public long ScheduleId { get; set; }
         public Schedule Schedule { get; set; }
     }
 }
